HBX-1977: Improve implementation of class ForeignKeyProcessor - Move ForeignKeyProcessor from 'org.hibernate.tool.internal.reveng' to 'org.hibernate.tool.internal.reveng.binder'
